Created basic fractal display framework
This is all pretty basic display type stuff; I just wanted to get a
working fractal going. The current design is that Main is going to
manage the fractals and the sounds, getting variables from the sound
player and sending them to the fractal to be visualized however the
fractal sees fit.

Right now there are 3 custom variables for the fractal to use that do
not have a declared purpose. When sound is implemented, then I can (I
hope!) plug values from the player into these. There could certainly be
more custom variables, but I don't see a reason to add them yet so I
won't bother.